rectif

+++ function_pointers/1-array_iterator.c
@@ -10,7 +10,7 @@
 #include "function_pointers.h"
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-size_t i;
+unsigned long i;
 for (i = 0; i < size; i++)
 action(array[i]);
 }
